requirements.txt
pyrogram
python-dotenv
tgcrypto
telegraph
python-decouple
pytgcalls